// src/rolls/ability-roll/ability-roll.types.ts
import type { AbilitySuccessLevelEnum } from "./ability-roll.defs";

export type Modifier = { description: string; value: number };

export type AbilityRollOptions = Partial<Roll.Options> & {
  naturalSkill: number; // Unmodified chance
  modifiers?: Modifier[];
  /** Alternative roll heading, used instead of flavor */
  heading?: string;
  abilityName?: string; // Skill name
  abilityType?: Item.SubType;
  abilityImg?: string; // Usually skill item image
  resultMessages?: Map<AbilitySuccessLevelEnum | undefined, string>; // Extra html to display in the roll
  speaker?: ChatMessage.SpeakerData;
  rollMode?: foundry.dice.Roll.Mode;
};